refactor: 멀티 모듈로 구성
