+++ idolu-product/src/main/java/com/idolu/product/ProductApplication.java
@@ -1,0 +1,13 @@
+package com.idolu.product;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ProductApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ProductApplication.class, args);
+    }
+
+}
