--- src/main/java/com/oneny/idoluproduct/IDolUProductApplication.java
@@ -1,13 +1,0 @@
-package com.oneny.idoluproduct;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class IDolUProductApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(IDolUProductApplication.class, args);
-	}
-
-}
